add bot to deployment

+++ apps/backend/src/models/read_log.js
@@ -22,7 +22,7 @@ module.exports = sequelize.define('ReadLog', {
     allowNull: false,
   },
   completion: {
-    type: DataTypes.TINYINT,
+    type: DataTypes.FLOAT,
     allowNull: false,
     validate: {
       max: 100,
